update Microservice package (#77)
* Initial commit

* Fix interfaces path

* Adding comments for api

* Added all the comments to public api

* Changed descriptions and added typedocs

* Fix doc style
