+++ .gitignore
@@ -5,6 +5,7 @@ es-bundle/
 lib/
 dist/
 flow-typed/
+doc/
 coverage/
 .npmrc
 yarn-error.log
