+++ .travis.yml
@@ -17,6 +17,7 @@ script:
   - yarn run build
   - yarn run lint
   - yarn run test
+  - yarn run doc
   - 'if [ "$TRAVIS_BRANCH" != "develop" ]; then bash scripts/publish.sh; fi'
 
 deploy:
