+++ packages/scalecube-microservice/src/ServiceCall/LocalCall.ts
@@ -1,7 +1,7 @@
 import { from, throwError, Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { throwErrorFromServiceCall } from '../helpers/utils';
-import { AddMessageToResponseOptions, InvokeMethodOptions, LocalCallOptions } from '../api/private/types';
+import { AddMessageToResponseOptions, InvokeMethodOptions, LocalCallOptions } from '../helpers/types';
 import {
   getAsyncModelMissmatch,
   getMethodNotFoundError,
