+++ packages/scalecube-microservice/src/api/LookUp.ts
@@ -1,0 +1,6 @@
+import { Endpoint, LookupOptions } from '.';
+
+/**
+ * The function that finds all the appropriate endpoints for a given criteria
+ */
+export type LookUp = (options: LookupOptions) => Endpoint[] | [];
