+++ packages/scalecube-microservice/src/api/Message.ts
@@ -1,0 +1,13 @@
+/**
+ * ServiceCall data
+ */
+export default interface Message {
+  /**
+   * The combination of serviceName and methodName: <serviceName/methodName>
+   */
+  qualifier: string;
+  /**
+   * Arguments of the invoked function
+   */
+  data: any;
+}
