+++ packages/scalecube-microservice/src/api/Microservices.ts
@@ -1,5 +1,11 @@
 import { MicroserviceOptions, Microservice } from '.';
 
+/**
+ * The factory for the creation of microservices containers
+ */
 export default interface Microservices {
+  /**
+   * The method for the creation of microservices containers
+   */
   create(microserviceOptions: MicroserviceOptions): Microservice;
 }
