--- packages/scalecube-microservice/src/api/public/CreateServiceCallOptions.ts
@@ -1,5 +1,0 @@
-import { Router } from '.';
-
-export default interface CreateServiceCallOptions {
-  router?: Router;
-}
