--- packages/scalecube-microservice/src/api/public/LookUp.ts
@@ -1,3 +1,0 @@
-import { Endpoint, LookupOptions } from './index';
-
-export type LookUp = (options: LookupOptions) => Endpoint[] | [];
