--- packages/scalecube-microservice/src/api/public/LookupOptions.ts
@@ -1,3 +1,0 @@
-export default interface LookupOptions {
-  qualifier: string;
-}
