--- packages/scalecube-microservice/src/api/public/Message.ts
@@ -1,4 +1,0 @@
-export default interface Message {
-  qualifier: string;
-  data: any;
-}
