--- packages/scalecube-microservice/src/api/public/Microservice.ts
@@ -1,9 +1,0 @@
-import { ProxyOptions, ServiceCall, CreateServiceCallOptions } from '.';
-
-export default interface Microservice {
-  destroy: () => null;
-
-  createProxy<T = any>(proxyOptions: ProxyOptions): T;
-
-  createServiceCall(options: CreateServiceCallOptions): ServiceCall;
-}
