--- packages/scalecube-microservice/src/api/public/MicroserviceOptions.ts
@@ -1,5 +1,0 @@
-import { Service } from '.';
-
-export default interface MicroserviceOptions {
-  services?: Service[];
-}
