--- packages/scalecube-microservice/src/api/public/ProxyOptions.ts
@@ -1,6 +1,0 @@
-import { ServiceDefinition, Router } from '.';
-
-export default interface ProxyOptions {
-  router?: Router;
-  serviceDefinition: ServiceDefinition;
-}
