--- packages/scalecube-microservice/src/api/public/RouteOptions.ts
@@ -1,6 +1,0 @@
-import { LookUp, Message } from '.';
-
-export default interface RouteOptions {
-  lookUp: LookUp;
-  message: Message;
-}
