--- packages/scalecube-microservice/src/api/public/Router.ts
@@ -1,5 +1,0 @@
-import { Endpoint, RouteOptions } from '.';
-
-export default interface Router {
-  route: (routeOptions: RouteOptions) => Endpoint | null;
-}
