--- packages/scalecube-microservice/src/api/public/Service.ts
@@ -1,6 +1,0 @@
-import { ServiceImplementation, ServiceDefinition } from '.';
-
-export default interface Service {
-  definition: ServiceDefinition;
-  reference: ServiceImplementation;
-}
