--- packages/scalecube-microservice/src/api/public/ServiceCall.ts
@@ -1,7 +1,0 @@
-import { Observable } from 'rxjs';
-import { Message } from '.';
-
-export default interface ServiceCall {
-  requestStream: (message: Message) => Observable<Message>;
-  requestResponse: (message: Message) => Promise<Message>;
-}
