+++ packages/scalecube-microservice/tests/implementations/ProxyFromMicroservice.spec.ts
@@ -3,8 +3,8 @@ import GreetingService, { greetingServiceDefinition } from '../mocks/GreetingSer
 import GreetingService2, { greetingServiceDefinition2 } from '../mocks/GreetingService2';
 import { Microservices, ASYNC_MODEL_TYPES } from '../../src';
 import { defaultRouter } from '../../src/Routers/default';
-import { ProxyOptions, Service, ServiceDefinition } from '../../src/api/public';
-import AsyncModel from '../../src/api/public/AsyncModel';
+import { ProxyOptions, Service, ServiceDefinition } from '../../src/api';
+import AsyncModel from '../../src/api/AsyncModel';
 import { expectWithFailNow } from '../helpers/utils';
 import {
   getAsyncModelMissmatch,
